ws: option to allow a refused websocket upgrade
Support a CURLWS_UPGRD_REFUSED_OK option that instructs libcurl not to error if the server refuses
to upgrade a connection to Websocket.  Also reuse cached http(s) connections if available for ws(s)
requests and return connections to the cache if the ws(s) connection is not upgraded.

In some circumstances a server may decide it should not or cannot upgrade to WebSocket and a client
may want to know that and continue without an error.  This is allowed according to the WebSocket
RFC, section 1.3 Opening Handshake: "Any status code other than 101 indicates that the WebSocket
handshake has not completed and that the semantics of HTTP still apply."  I have a specific existing
architecture I'd like to integrate libcurl into that requires this to work.

I understand the reasoning behind the documented behavior to fail the request if a ws(s):// scheme
is requested and the server doesn't respond with a 101 switching protocol response.  However
supporting this option will give the user the ability to allow it and be smarter about connection
reuse.

If an upgrade is refused and the new option is set, the http(s) connection can be returned to the
cache for reuse if left intact.  In addition, an http(s) connection can now be used off the cache if
the url otherwise matches correctly when a WebSocket upgrade is needed.  One thing to consider might
be hiding this behavior behind another option.

Full disclosures: I used AI to help me write the unit test but not any of the main library code.